envsetup: migrate container setup to ansible playbook
now the container can be updated via a second run of env.sh
this triggers ansible again and the new playbook will update
the container. initial oe build need a long time to build
it's toolchain therefore an internal update mechanism is
more efficient than recreating the container all the time

+++ env.sh
@@ -1,5 +1,7 @@
 #!/bin/bash -e
 
+NAMESERVER="8.8.8.8"
+
 die() {
   echo "${1}" && exit 1
 }
@@ -29,46 +31,33 @@ else
 fi
 
 if [ "${2}" = "" ]; then
-  DEVDIR=$(pwd)
+  DEVDIR=$(dirname $(dirname $(realpath ${0})))
 else
   DEVDIR=${2}
 fi
 
-if ! [ -d "${DEVDIR}/meta-fnordpipe" ]; then
-  die "missing repo in DEVDIR"
+if ! lxc-info -n "${INSTANCE}" &> /dev/null; then
+  export LXCCONFIG=$(mktemp)
+  export DEBIAN_FRONTEND=noninteractive
+
+  lxccfg ${LXCCONFIG}
+  lxc-create -n "${INSTANCE}" -f ${LXCCONFIG} -t ubuntu -- -r xenial
+  rm ${LXCCONFIG}
+  lxc-start -n "${INSTANCE}"
+
+  lxcrun 'rm /etc/resolv.conf'
+  lxcrun "echo nameserver ${NAMESERVER} | tee /etc/resolv.conf"
+
+  # create user
+  lxcrun 'userdel ubuntu'
+  LXCUID=$(ls -an ${DEVDIR} | sed -n 2p | awk '{ print $3 }')
+  LXCGID=$(ls -an ${DEVDIR} | sed -n 2p | awk '{ print $4 }')
+  lxcrun "groupadd -g ${LXCGID} user"
+  lxcrun "useradd -m -g ${LXCGID} -u ${LXCUID} user"
+
+  lxcrun 'apt-get update'
+  lxcrun 'apt-get install -y ansible aptitude'
 fi
 
-lxc-info -n "${INSTANCE}" &> /dev/null && die "already exists"
-
-export LXCCONFIG=$(mktemp)
-export DEBIAN_FRONTEND=noninteractive
-
-lxccfg ${LXCCONFIG}
-lxc-create -n "${INSTANCE}" -f ${LXCCONFIG} -t ubuntu -- -r xenial
-rm ${LXCCONFIG}
-lxc-start -n "${INSTANCE}"
-
-lxcrun 'rm /etc/resolv.conf'
-lxcrun 'echo nameserver 8.8.8.8 | tee /etc/resolv.conf'
-lxcrun 'apt-get update'
-
-# create user
-LXCUID=$(ls -an ${DEVDIR} | sed -n 2p | awk '{ print $3 }')
-LXCGID=$(ls -an ${DEVDIR} | sed -n 2p | awk '{ print $4 }')
-lxcrun 'userdel ubuntu'
-lxcrun "groupadd -g ${LXCGID} user"
-lxcrun "useradd -m -g ${LXCGID} -u ${LXCUID} user"
-lxcrun 'ln -s /usr/local/development ~user/development'
-lxcrun 'ln -s /var/lib/oe ~user/oe'
-
-lxcrun "usermod -p '\$6\$VasvKpIc\$WjeeFZ9e8nl0FGdUHDWwrFC.6WMXtyPA5kIplZL6.SlVgcUxOf/sWthnUutR4EAmmhZdjhQZIfUuaKCI9bPiM0' root"
-
-# install oe dependencies
-lxcrun 'apt-get install -y cpio python3 gawk wget git-core diffstat unzip texinfo gcc-multilib build-essential chrpath libsdl1.2-dev'
-lxcrun 'install -m 0750 -o user -g user -d /var/lib/oe'
-lxcrun '/bin/su -l user -c "git clone git://git.openembedded.org/openembedded-core /var/lib/oe"'
-lxcrun '/bin/su -l user -c "git clone git://git.openembedded.org/bitbake /var/lib/oe/bitbake"'
-lxcrun '/bin/su -l user -c "cd /var/lib/oe && git checkout pyro"'
-lxcrun '/bin/su -l user -c "cd /var/lib/oe/bitbake && git checkout 1.34"'
-lxcrun '/bin/su -l user -c "ln -s /usr/local/development/meta-fnordpipe /var/lib/oe/meta-fnordpipe"'
-lxcrun '/bin/su -l user -c "cd /var/lib/oe && export TEMPLATECONF=meta-fnordpipe/conf/samples && source ./oe-init-build-env"'
+ANSIBLEDIR="/usr/local/development/$(basename $(dirname $(realpath ${0})))/.env"
+lxcrun "/bin/su -l -c 'cd ${ANSIBLEDIR} && ansible-playbook -i inventory/hosts.ini prepare.yml'"
